+++ RenderEngine/Source/Core/Object/Components/MeshRenderer/Mesh.cpp
@@ -2,7 +2,7 @@
 
 using namespace RenderEngine::Core;
 
-void Mesh::InitializeMesh(std::vector<Vertex> _vertices, std::vector<uint32_t> _indices, Mesh* _output)
+void Mesh::InitializeMesh(std::vector<Vertex> _vertices, std::vector<uint16_t> _indices, Mesh* _output)
 {
 	_output->indices = _indices;
 	_output->vertices = _vertices;
@@ -13,7 +13,7 @@ std::vector<Vertex> Mesh::GetVertices() const
 	return vertices;
 }
 
-std::vector<uint32_t> Mesh::GetIndices() const
+std::vector<uint16_t> Mesh::GetIndices() const
 {
 	return indices;
 }
